refactor(exception): 패키지 경로 racingcar.exception으로 이동
기존의 common 패키지에서 exception 패키지로 이동하여 예외 관련 코드를 명확히 구분하고, 코드 가독성과 관리 용이성을 개선하였습니다.
