+++ src/test/java/racingcar/dto/AttemptCountTest.java
@@ -1,13 +1,13 @@
 package racingcar.dto;
 
 import static org.assertj.core.api.Assertions.*;
-import static racingcar.common.exception.ErrorMessage.*;
+import static racingcar.exception.ErrorMessage.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.common.exception.RacingCarException;
+import racingcar.exception.RacingCarException;
 
 
 class AttemptCountTest {
